Fix for inferring the same dependency twice.
When using `module.declare(function (require, exports, module) { ... })`, if you `require`'d the same module twice, the dependency inference algorithm would infer it twice and then try to provide it twice. This eliminates those extra calls to `module.provide`, which we must remember could be in the control of a provider plug-in that isn't as well-equipped to deal with this behavior.